bug fixes and new features
in "Create annonymous patiet" form - combos of barrier and hospital are now sorted
now it is possible to edit ride details and then change driver.

added ID number to volunteers.

all of the annonymous patients are now displayed as "patient".

it is now possible to change the time of a ride even if a driver is assigned to ride.

started to work on the LastModified of a ride.