manageRidePats.html + login
small fix, commented out a single row in refreshTable()
which cause an issue, the issue was wrong value of hoisted parameter i which was incompetible with the arr_rides.
after carefull consideration it seems to be a redundant row. (I commented it out and not delete in act of catious)